stable_diffusion_inference-main/requirements.txt
opencv-python~=4.6.0.66
Pillow
pudb==2019.2
imageio
imageio-ffmpeg
pytorch-lightning
omegaconf
einops==0.3.0
transformers>4.19.2
open-clip-torch==2.7.0
torch
numpy
torchvision
kornia
scipy
setuptools
lightning
